adjust datatype to be long

+++ 0x05-pointers_arrays_strings/6-puts2.c
@@ -5,7 +5,7 @@
  */
 void puts2(char *str)
 {
-	int pos;
+	long int pos;
 
 	for (pos = 0; *(str + pos) != '\0'; pos += 2)
 	{
